+++ cmd/template/cluster/runner.go
@@ -125,6 +125,11 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 		if err != nil {
 			return microerror.Mask(err)
 		}
+	case key.ProviderOpenStack:
+		err = provider.WriteCAPOTemplate(ctx, c.K8sClient, output, config)
+		if err != nil {
+			return microerror.Mask(err)
+		}
 	case key.ProviderVsphere:
 		err = provider.WriteCAPVTemplate(ctx, c.K8sClient, output, config)
 		if err != nil {