+++ internal/key/provider.go
@@ -4,5 +4,6 @@ const (
 	ProviderAWS     = "aws"
 	ProviderAzure   = "azure"
 	ProviderKVM     = "kvm"
+	ProviderOpenStack = "openstack"
 	ProviderVsphere = "vsphere"
 )